Split AutoUpdater into extension files
Extract large AutoUpdater responsibilities into focused extensions. Adds AutoUpdater+BackgroundScheduler.swift, AutoUpdater+Install.swift and AutoUpdater+InstallAndRelaunch.swift and removes the background-scheduling and install/relaunch blocks from AutoUpdater.swift. Preserves existing logic, documentation and shutdown/install flow (unzip→replaceItem→open -n→NSApp.terminate) while keeping the main AutoUpdater file smaller and more modular.
